[UPDATE] 경로수정 -마이페이지 경로수정
