+++ src/main/webapp/WEB-INF/views/member/mypage.jsp
@@ -63,7 +63,7 @@
     </div>
     <div class="mypage-container mypage-container-myinfo">
       <sec:authorize access="isAuthenticated()">
-        <a href="/member/user/myInfo?id=<sec:authentication property="name" />">내 정보</a>
+        <a href="/member/myInfo?id=<sec:authentication property="name" />">내 정보</a>
       </sec:authorize>
     </div>
     <div class="mypage-container">
